CountdownSeconds component has its own file
